programacion1/servicio/mysite/requirements.txt
asgiref==3.3.4
Django==3.2.4
django-mysql==3.12.0
djangorestframework==3.12.4
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
